Split DragGroup into its own file.
